Comments, use deque to store continuations in Delegate
